+++ docs/CHANGELOG.rst
@@ -18,6 +18,8 @@ Added
 Changed
 -------
 * Use values instead of abbreviations for genome sizes in chip_seq
+* Utility functions return only one element instead of list when thay
+  are run on a ``Data`` object
 
 
 ==================